+++ src/cljs/lambdacd/db.cljs
@@ -11,7 +11,8 @@
    :connection-state :lost
    :displayed-build-number nil
    :raw-step-results-visible false
-   :expanded-step-ids #{}})
+   :expanded-step-ids #{}
+   :expand-active false})
 
 (defn initialize-db-handler [_ _]
   default-db)
@@ -75,9 +76,16 @@
 (defn current-step-result-subscription [db _]
   (reaction (state/find-by-step-id (:pipeline-state @db) (:step-id @db))))
 
+(defn- expand-active? [db step-id]
+  (and
+    (:expand-active db)
+    (state/is-active? (state/find-by-step-id (:pipeline-state db) step-id))))
+
 (defn step-expanded-subscription [db-atom [_ step-id]]
   (reaction (let [db @db-atom]
-                (contains? (:expanded-step-ids db) step-id))))
+              (or
+                (contains? (:expanded-step-ids db) step-id)
+                (expand-active? db step-id)))))
 
 (defn all-step-ids [{state :pipeline-state}]
   (->> state
@@ -104,6 +112,12 @@
 (defn set-all-collapsed-handler [db _]
   (assoc db :expanded-step-ids #{}))
 
+(defn toggle-expand-active-handler [db _]
+  (update db :expand-active not))
+
+(defn expand-active-active-subscription [db _]
+  (reaction (:expand-active @db)))
+
 (re-frame/register-handler ::history-updated history-updated-handler)
 (re-frame/register-handler ::initialize-db initialize-db-handler)
 (re-frame/register-handler ::pipeline-state-updated pipeline-state-updated-handler)
@@ -114,6 +128,7 @@
 (re-frame/register-handler ::toggle-step-expanded toggle-step-expanded)
 (re-frame/register-handler ::set-all-expanded set-all-expanded-handler)
 (re-frame/register-handler ::set-all-collapsed set-all-collapsed-handler)
+(re-frame/register-handler ::toggle-expand-active toggle-expand-active-handler)
 
 (re-frame/register-sub ::history history-subscription)
 (re-frame/register-sub ::pipeline-state pipeline-state-subscription)
@@ -125,3 +140,4 @@
 (re-frame/register-sub ::step-expanded? step-expanded-subscription)
 (re-frame/register-sub ::all-expanded? all-expanded-subscription)
 (re-frame/register-sub ::all-collapsed? all-collapsed-subscription)
+(re-frame/register-sub ::expand-active-active? expand-active-active-subscription)